requirements.txt
fastapi
uvicorn[standard]
python-multipart
transformers
torch
torchvision
pillow
